add, fix : add and fix files
Makefile, ft_lstmap, ft_memchr, libft

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: wooshin <wooshin@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/11 07:33:29 by wooshin           #+#    #+#             */
-/*   Updated: 2022/07/12 03:57:03 by wooshin          ###   ########.fr       */
+/*   Updated: 2022/08/18 04:58:32 by wooshin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
